+++ R/regexpr.R
@@ -1,16 +1,19 @@
 #' @rdname regexpr-style
+#' @family regex-utils
 #' @export
 set_PCRE <- function() {
     options(regexpr_use_perl = TRUE)
 }
 
 #' @rdname regexpr-style
+#' @family regex-utils
 #' @export
 set_TRE <- function() {
     options(regexpr_use_perl = FALSE)
 }
 
 #' @rdname regexpr-style
+#' @family regex-utils
 #' @export
 use_perl <- function() {
     res <- getOption("regexpr_use_perl", default = auto_set_regexpr_style())
@@ -18,23 +21,20 @@ use_perl <- function() {
 }
 
 
-#' switch regular expression style (PCRE vs TRE)
+#' Switch regular expression style (PCRE vs TRE)
+#' @family regex-utils
 #'
-#' The `set_regexpr_style()` allows user to specify which style to be used, 
-#' while the `auto_set_regexpr_style()` automatically set the style depdending on 
-#' the operating system (TRE for Windows and PCRE for other OSs (Linux and Mac)).
-#' 
-#' `set_PCRE()` force to use PCRE style while `set_TRE()` force to use TRE. 
-#' 
-#' Note that all these functions are not change the behavior of `gsub()` and `regexpr()`. 
-#' The functions are just set a global option to store the user's choice of whether using `perl = TRUE`.
+#' - `set_regexpr_style()` selects the style explicitly.
+#' - `auto_set_regexpr_style()` chooses based on OS (TRE on Windows; PCRE elsewhere).
+#' - `set_PCRE()` and `set_TRE()` force the style.
 #'
-#' Users can access the option via `use_perl()` and pass the return value to `gusb()` or `regexpr()` to specify the style in use.
-#' 
+#' These functions do not change the behavior of `gsub()`/`regexpr()` directly.
+#' They set a global option that you can read via `use_perl()` and pass to `gsub()`/`regexpr()`.
+#'
 #' @rdname regexpr-style
 #' @param style one of 'PCRE' or 'TRE'
-#' @return logical value of whether use perl
-#' @references <https://stackoverflow.com/questions/47240375/regular-expressions-in-base-r-perl-true-vs-the-default-pcre-vs-tre>
+#' @return Logical indicating whether to use `perl`
+#' @references https://stackoverflow.com/questions/47240375/regular-expressions-in-base-r-perl-true-vs-the-default-pcre-vs-tre
 #' @export
 #' @author Guangchuang Yu
 set_regexpr_style <- function(style) {