+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(yulab.utils)
+
+test_check("yulab.utils")